Add Azure function generating tips

+++ TakeAIMeal.Common.Services/Interfaces/ITextGeneratorService.cs
@@ -10,9 +10,9 @@
         /// </summary>
         /// <param name="prompt">The input text prompt to generate text from.</param>
         /// <param name="completionCount">The number of text completions to generate (default is 1).</param>
-        /// <param name="temperature">The sampling temperature to use for generating text (default is 0.2f).</param>
+        /// <param name="temperature">The sampling temperature to use for generating text (default is 0.2).</param>
         /// <param name="maxTokens">The maximum number of tokens (words or subwords) to generate in each text completion (default is 1024).</param>
         /// <returns>A task that represents the asynchronous operation. The task result contains a string collection that represents the generated text.</returns>
-        Task<ICollection<string>> GenerateText(string prompt, int completionCount = 1, float temperature = 0.2f, int maxTokens = 1024);
+        Task<ICollection<string>> GenerateText(string prompt, int completionCount = 1, double temperature = 0.2, int maxTokens = 1024);
     }
 }
